fix(treewide): inconsistent use of unsigned long vs size_t
On *most* platforms, these are the same. However, there are certain platforms (e.g. mingw on windows 32-bit) where unsigned long is less than 64 bits (namely, 32 bits) while a size_t (which is what the cpp header accepts) is 64 bits.

This technically breaks ABI compatibility on mingw 32-bit specifically, but this won't actually have broken anything because any program that used the previous behaviour was already broken. On all other currently supported platforms, like with the previous change, size_t and unsigned long happen to be the same, and so we maintain ABI and source backwards compatibility there.

+++ include/lsl_cpp.h
@@ -629,7 +629,7 @@ public:
 		const std::vector<T> &buffer, double timestamp = 0.0, bool pushthrough = true) {
 		if (!buffer.empty())
 			push_chunk_multiplexed(
-				buffer.data(), static_cast<unsigned long>(buffer.size()), timestamp, pushthrough);
+				buffer.data(), static_cast<size_t>(buffer.size()), timestamp, pushthrough);
 	}
 
 	/** Push a chunk of multiplexed data into the outlet. One timestamp per sample is provided.
@@ -647,7 +647,7 @@ public:
 		const std::vector<double> &timestamps, bool pushthrough = true) {
 		if (!buffer.empty() && !timestamps.empty())
 			push_chunk_multiplexed(
-				buffer.data(), static_cast<unsigned long>(buffer.size()), timestamps.data(), pushthrough);
+				buffer.data(), static_cast<size_t>(buffer.size()), timestamps.data(), pushthrough);
 	}
 
 	/** Push a chunk of multiplexed samples into the outlet. Single timestamp provided.
@@ -665,32 +665,32 @@ public:
 	 */
 	void push_chunk_multiplexed(const float *buffer, std::size_t buffer_elements,
 		double timestamp = 0.0, bool pushthrough = true) {
-		lsl_push_chunk_ftp(obj.get(), buffer, static_cast<unsigned long>(buffer_elements), timestamp, pushthrough);
+		lsl_push_chunk_ftp(obj.get(), buffer, static_cast<size_t>(buffer_elements), timestamp, pushthrough);
 	}
 	void push_chunk_multiplexed(const double *buffer, std::size_t buffer_elements,
 		double timestamp = 0.0, bool pushthrough = true) {
-		lsl_push_chunk_dtp(obj.get(), buffer, static_cast<unsigned long>(buffer_elements), timestamp, pushthrough);
+		lsl_push_chunk_dtp(obj.get(), buffer, static_cast<size_t>(buffer_elements), timestamp, pushthrough);
 	}
 	void push_chunk_multiplexed(const int64_t *buffer, std::size_t buffer_elements,
 		double timestamp = 0.0, bool pushthrough = true) {
-		lsl_push_chunk_ltp(obj.get(), buffer, static_cast<unsigned long>(buffer_elements), timestamp, pushthrough);
+		lsl_push_chunk_ltp(obj.get(), buffer, static_cast<size_t>(buffer_elements), timestamp, pushthrough);
 	}
 	void push_chunk_multiplexed(const int32_t *buffer, std::size_t buffer_elements,
 		double timestamp = 0.0, bool pushthrough = true) {
-		lsl_push_chunk_itp(obj.get(), buffer, static_cast<unsigned long>(buffer_elements), timestamp, pushthrough);
+		lsl_push_chunk_itp(obj.get(), buffer, static_cast<size_t>(buffer_elements), timestamp, pushthrough);
 	}
 	void push_chunk_multiplexed(const int16_t *buffer, std::size_t buffer_elements,
 		double timestamp = 0.0, bool pushthrough = true) {
-		lsl_push_chunk_stp(obj.get(), buffer, static_cast<unsigned long>(buffer_elements), timestamp, pushthrough);
+		lsl_push_chunk_stp(obj.get(), buffer, static_cast<size_t>(buffer_elements), timestamp, pushthrough);
 	}
 	void push_chunk_multiplexed(const int8_t *buffer, std::size_t buffer_elements,
 		double timestamp = 0.0, bool pushthrough = true) {
-		lsl_push_chunk_ctp(obj.get(), buffer, static_cast<unsigned long>(buffer_elements), timestamp, pushthrough);
+		lsl_push_chunk_ctp(obj.get(), buffer, static_cast<size_t>(buffer_elements), timestamp, pushthrough);
 	}
 	void push_chunk_multiplexed(const char *buffer, std::size_t buffer_elements,
 		double timestamp = 0.0, bool pushthrough = true) {
 		lsl_push_chunk_ctp(obj.get(), reinterpret_cast<const int8_t *>(buffer),
-			static_cast<unsigned long>(buffer_elements), timestamp, pushthrough);
+			static_cast<size_t>(buffer_elements), timestamp, pushthrough);
 	}
 	void push_chunk_multiplexed(const std::string *buffer, std::size_t buffer_elements,
 		double timestamp = 0.0, bool pushthrough = true) {
@@ -702,7 +702,7 @@ public:
 				lengths[k] = (uint32_t)buffer[k].size();
 			}
 			lsl_push_chunk_buftp(obj.get(), pointers.data(), lengths.data(),
-				static_cast<unsigned long>(buffer_elements), timestamp, pushthrough);
+				static_cast<size_t>(buffer_elements), timestamp, pushthrough);
 		}
 	}
 
@@ -721,38 +721,38 @@ public:
 	 */
 	void push_chunk_multiplexed(const float *data_buffer, const double *timestamp_buffer,
 		std::size_t data_buffer_elements, bool pushthrough = true) {
-		lsl_push_chunk_ftnp(obj.get(), data_buffer, static_cast<unsigned long>(data_buffer_elements),
+		lsl_push_chunk_ftnp(obj.get(), data_buffer, static_cast<size_t>(data_buffer_elements),
 			(timestamp_buffer), pushthrough);
 	}
 	void push_chunk_multiplexed(const double *data_buffer, const double *timestamp_buffer,
 		std::size_t data_buffer_elements, bool pushthrough = true) {
-		lsl_push_chunk_dtnp(obj.get(), data_buffer, static_cast<unsigned long>(data_buffer_elements),
+		lsl_push_chunk_dtnp(obj.get(), data_buffer, static_cast<size_t>(data_buffer_elements),
 			(timestamp_buffer), pushthrough);
 	}
 	void push_chunk_multiplexed(const int64_t *data_buffer, const double *timestamp_buffer,
 		std::size_t data_buffer_elements, bool pushthrough = true) {
-		lsl_push_chunk_ltnp(obj.get(), data_buffer, static_cast<unsigned long>(data_buffer_elements),
+		lsl_push_chunk_ltnp(obj.get(), data_buffer, static_cast<size_t>(data_buffer_elements),
 			(timestamp_buffer), pushthrough);
 	}
 	void push_chunk_multiplexed(const int32_t *data_buffer, const double *timestamp_buffer,
 		std::size_t data_buffer_elements, bool pushthrough = true) {
-		lsl_push_chunk_itnp(obj.get(), data_buffer, static_cast<unsigned long>(data_buffer_elements),
+		lsl_push_chunk_itnp(obj.get(), data_buffer, static_cast<size_t>(data_buffer_elements),
 			(timestamp_buffer), pushthrough);
 	}
 	void push_chunk_multiplexed(const int16_t *data_buffer, const double *timestamp_buffer,
 		std::size_t data_buffer_elements, bool pushthrough = true) {
-		lsl_push_chunk_stnp(obj.get(), data_buffer, static_cast<unsigned long>(data_buffer_elements),
+		lsl_push_chunk_stnp(obj.get(), data_buffer, static_cast<size_t>(data_buffer_elements),
 			(timestamp_buffer), pushthrough);
 	}
 	void push_chunk_multiplexed(const int8_t *data_buffer, const double *timestamp_buffer,
 		std::size_t data_buffer_elements, bool pushthrough = true) {
-		lsl_push_chunk_ctnp(obj.get(), data_buffer, static_cast<unsigned long>(data_buffer_elements),
+		lsl_push_chunk_ctnp(obj.get(), data_buffer, static_cast<size_t>(data_buffer_elements),
 			(timestamp_buffer), pushthrough);
 	}
 	void push_chunk_multiplexed(const char *data_buffer, const double *timestamp_buffer,
 		std::size_t data_buffer_elements, bool pushthrough = true) {
 		lsl_push_chunk_ctnp(obj.get(), reinterpret_cast<const int8_t *>(data_buffer),
-			static_cast<unsigned long>(data_buffer_elements), (timestamp_buffer), pushthrough);
+			static_cast<size_t>(data_buffer_elements), (timestamp_buffer), pushthrough);
 	}
 
 	void push_chunk_multiplexed(const std::string *data_buffer, const double *timestamp_buffer,
@@ -765,7 +765,7 @@ public:
 				lengths[k] = (uint32_t)data_buffer[k].size();
 			}
 			lsl_push_chunk_buftnp(obj.get(), pointers.data(), lengths.data(),
-				static_cast<unsigned long>(data_buffer_elements), timestamp_buffer, pushthrough);
+				static_cast<size_t>(data_buffer_elements), timestamp_buffer, pushthrough);
 		}
 	}
 
@@ -1294,7 +1294,7 @@ public:
 		double timeout = 0.0) {
 		int32_t ec = 0;
 		std::size_t res = lsl_pull_chunk_f(obj.get(), data_buffer, timestamp_buffer,
-			(unsigned long)data_buffer_elements, (unsigned long)timestamp_buffer_elements, timeout,
+			(size_t)data_buffer_elements, (size_t)timestamp_buffer_elements, timeout,
 			&ec);
 		check_error(ec);
 		return res;
@@ -1304,7 +1304,7 @@ public:
 		double timeout = 0.0) {
 		int32_t ec = 0;
 		std::size_t res = lsl_pull_chunk_d(obj.get(), data_buffer, timestamp_buffer,
-			(unsigned long)data_buffer_elements, (unsigned long)timestamp_buffer_elements, timeout,
+			(size_t)data_buffer_elements, (size_t)timestamp_buffer_elements, timeout,
 			&ec);
 		check_error(ec);
 		return res;
@@ -1314,7 +1314,7 @@ public:
 		double timeout = 0.0) {
 		int32_t ec = 0;
 		std::size_t res = lsl_pull_chunk_l(obj.get(), data_buffer, timestamp_buffer,
-			(unsigned long)data_buffer_elements, (unsigned long)timestamp_buffer_elements, timeout,
+			(size_t)data_buffer_elements, (size_t)timestamp_buffer_elements, timeout,
 			&ec);
 		check_error(ec);
 		return res;
@@ -1324,7 +1324,7 @@ public:
 		double timeout = 0.0) {
 		int32_t ec = 0;
 		std::size_t res = lsl_pull_chunk_i(obj.get(), data_buffer, timestamp_buffer,
-			(unsigned long)data_buffer_elements, (unsigned long)timestamp_buffer_elements, timeout,
+			(size_t)data_buffer_elements, (size_t)timestamp_buffer_elements, timeout,
 			&ec);
 		check_error(ec);
 		return res;
@@ -1334,7 +1334,7 @@ public:
 		double timeout = 0.0) {
 		int32_t ec = 0;
 		std::size_t res = lsl_pull_chunk_s(obj.get(), data_buffer, timestamp_buffer,
-			(unsigned long)data_buffer_elements, (unsigned long)timestamp_buffer_elements, timeout,
+			(size_t)data_buffer_elements, (size_t)timestamp_buffer_elements, timeout,
 			&ec);
 		check_error(ec);
 		return res;
@@ -1344,7 +1344,7 @@ public:
 		double timeout = 0.0) {
 		int32_t ec = 0;
 		std::size_t res = lsl_pull_chunk_c(obj.get(), data_buffer, timestamp_buffer,
-			static_cast<unsigned long>(data_buffer_elements), static_cast<unsigned long>(timestamp_buffer_elements), timeout,
+			static_cast<size_t>(data_buffer_elements), static_cast<size_t>(timestamp_buffer_elements), timeout,
 			&ec);
 		check_error(ec);
 		return res;
@@ -1367,8 +1367,8 @@ public:
 			std::vector<char *> result_strings(data_buffer_elements);
 			std::vector<uint32_t> result_lengths(data_buffer_elements);
 			std::size_t num = lsl_pull_chunk_buf(obj.get(), result_strings.data(), result_lengths.data(),
-				timestamp_buffer, static_cast<unsigned long>(data_buffer_elements),
-				static_cast<unsigned long>(timestamp_buffer_elements), timeout, &ec);
+				timestamp_buffer, static_cast<size_t>(data_buffer_elements),
+				static_cast<size_t>(timestamp_buffer_elements), timeout, &ec);
 			check_error(ec);
 			for (std::size_t k = 0; k < num; k++) {
 				data_buffer[k].assign(result_strings[k], result_lengths[k]);
